rename theme.styl
