+++ src/components/action-bar/index.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { translate } from 'react-i18next';
-import style from './theme.styl';
+import style from './style.styl';
 
 export default (props) => (
         <div id="action-bar" className={style['actions-bar']}>
