+++ src/components/switch/style.styl
