+++ src/components/tabs/style.styl
